[todoApp]スタイル読み込まなくなったけどjs分離できた！
